P: New is not go-style
S: ... it looks like C method, however in Go we can be smarted and
type-safe at the same time. So introduce `ZyreOption` family of
functions, which can set some parts for Zyre during `New` phase
