+++ README.md
@@ -8,24 +8,11 @@ A golang interface to the [Zyre v2.0](http://github.com/zeromq/zyre) API.
 1. Most of methods wrapped 1:1
 2. Missing - `zyre_whisper` and `zyre_shout`
 3. Should rename `Shouts` to `ShoutString`
-4. Options can be similear to socket
-```go
-    func NewSock(t int, options ...SockOption) *Sock {
-
-// SockOption is a type for setting options on the underlying ZeroMQ socket
-type SockOption func(*Sock)
-
-// SetOption accepts a SockOption and uses it to set an option on
-// the underlying ZeroMQ socket
-func (s *Sock) SetOption(o SockOption) {
-	o(s)
-}
-
-IOW one should set most interesting stuff at New and then call Start + Join + ...
-```
-5. Examples
-6. Wrap zyre_socket??? - IOW this does not make a sense for golang, much easier
-   is to call Recv from gorutine and pass the result to channel
+4. Examples
+5. Options
+    - right now there are `Zyre.SetVerbose` and `SetVerbose` methods, maybe keep booth
+    - move gossip configuration to extra code
+      NewGossip(name, endpoint, bind, connect)
 
 # Install
 ## Dependencies
@@ -59,7 +46,10 @@ func main() {
 	flag.StringVar(&name, "name", "", "node name")
 	flag.Parse()
 
-	node := zyre.New(name)
+	node := zyre.New(
+        name,
+        zyre.SetHeader("foo", "bar"),
+        )
 	err := node.Start()
 	if err != nil {
 		panic(err)
